requirements.txt
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
monai==1.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
pillow==11.0.0
pynrrd==1.1.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==70.2.0
SimpleITK==2.4.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
triton==3.2.0
typing_extensions==4.12.2
